% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_block.R
\name{make_block}
\alias{make_block}
\title{Making time-lagged block}
\usage{
make_block(X, lib = c(1, NROW(X)), lag = 0, group = NULL)
}
\arguments{
\item{X}{a data.frame or matrix where each column is a time series.}

\item{lib}{a two-column matrix (or two-element vector) where each row specifies the
first and last indices of time series to use for attractor reconstruction.}

\item{lag}{the time-lag. Must be an integer.}

\item{group}{the name or column index of a group variable, which specifies different
groups of time series data.}
}
\value{
A data.frame where each column represents a time-lagged variable.
}
\description{
\code{make_block} makes time-lagged block.
}
\examples{
# simulate logistic map
tl <- 400  # time length
x <- y <- rep(NA, tl)
x[1] <- 0.4
y[1] <- 0.2
for (t in 1:(tl - 1)) {  # causality : x -> y
    x[t+1] = x[t] * (3.8 - 3.8 * x[t] - 0.0 * y[t])
    y[t+1] = y[t] * (3.5 - 3.5 * y[t] - 0.1 * x[t])
}
block <- data.frame(t=1:tl, x=x, y=y)

head(make_block(block[-1], lag= 0))
head(make_block(block[-1], lag= 1))
head(make_block(block[-1], lag=-1))

}
