% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_uic.R
\name{marginal_uic}
\alias{marginal_uic}
\title{Perform simplex projection and compute marginal unified information-theoretic causality}
\usage{
marginal_uic(
  block,
  lib = c(1, NROW(block)),
  pred = lib,
  lib_var = 1,
  tar_var = 2,
  cond_var = NULL,
  norm = 2,
  E = 1,
  tau = 1,
  tp = 0,
  nn = "e+1",
  n_boot = 2000,
  scaling = c("neighbor", "velocity", "no_scale"),
  exclusion_radius = NULL,
  epsilon = NULL,
  is_naive = FALSE
)
}
\arguments{
\item{block}{a data.frame or matrix where each column is a time series.}

\item{lib}{a two-column matrix (or two-element vector) where each row specifies
the first and last indices of time series to use for attractor reconstruction.}

\item{pred}{a two-column matrix (or two-element vector) where each row specifies
the first and last indices of time series to use for model predictions.}

\item{lib_var}{the names or column indices of library variables.
The specified variables are used as explanatory variables with time-delay embedding.}

\item{tar_var}{the name or column index of a target variable.
The specified variable is used as response variables.}

\item{cond_var}{the names or column indeices of condition data.
The specified variables are used as explanatory variables (without time-delay embedding).}

\item{norm}{the power of Lp distance to use. See Details.}

\item{E}{the embedding dimension to use for time-delay embedding.}

\item{tau}{the time-lag to use for time-delay embedding.}

\item{tp}{the time index to predict.}

\item{nn}{the number of nearest neighbors to use.
If \code{nn = "e+1"}, \code{nn} is set as \code{E} + 1.}

\item{n_boot}{the number of bootstraps to use for computing p-value.}

\item{scaling}{the method for local scaling of distance matrix. See Details.}

\item{exclusion_radius}{the filtering to exclude nearest neighbors if their time index is within exclusion radius.}

\item{epsilon}{the filtering to exclude nearest neighbors if epsilon is farther than their distances.}

\item{is_naive}{specifies whether the rEDM-style estimator is used.}
}
\value{
A data.frame where each row represents model statistics computed from a parameter set.
\tabular{ll}{
\code{E_best}   \tab \code{:} embedding dimension, which minimize RMSE for simplex projection \cr
\code{tau_best} \tab \code{:} time-lag, which minimize RMSE for simplex projection \cr
\code{tp}       \tab \code{:} time prediction horizon \cr
\code{nn_best}  \tab \code{:} number of nearest neighbors, which minimize RMSE for simplex projection \cr
\code{te}       \tab \code{:} transfer entropy \cr
\code{pval}     \tab \code{:} bootstrap p-value for te > 0 \cr
}

\code{te} is transfer entropy based on the unified information-theoretic causality test:
\deqn{
\sum_{t} log p(y_{t+tp} | x_{t}     , x_{t- \tau}, \ldots, x_{t-(E-1)\tau}, z_{t}) -
         log p(y_{t+tp} | x_{t-\tau}, x_{t-2\tau}, \ldots, x_{t-(E-1)\tau}, z_{t})
}
where \eqn{x} is library, \eqn{y} is target and \eqn{z} is condition.
}
\description{
\code{marginal_uic} returns model statistics computed from given multiple time series
based on simplex projection and cross mapping. This function computes UIC using model
averaging technique (and marginalizing \code{E} and \code{tau}). Thus, the users do not
have to determine the optimal \code{E} and \code{tau}.
}
\examples{
## simulate logistic map
tl <- 400  # time length
x <- y <- rep(NA, tl)
x[1] <- 0.4
y[1] <- 0.2
for (t in 1:(tl - 1)) {  # causality : x -> y
    x[t+1] = x[t] * (3.8 - 3.8 * x[t] - 0.0 * y[t])
    y[t+1] = y[t] * (3.5 - 3.5 * y[t] - 0.1 * x[t])
}
block <- data.frame(t = 1:tl, x = x, y = y)

## UIC
op0 = marginal_uic(block, lib_var = "x", tar_var = "y", E = 1:10, tau = 1:3, tp = -4:4,
                   n_boot = 2000)
op1 = marginal_uic(block, lib_var = "y", tar_var = "x", E = 1:10, tau = 1:3, tp = -4:4,
                   n_boot = 2000)
par(mfrow = c(2, 1))
with(op0, plot(tp, te, type = "l"))
with(op0[op0$pval < 0.05,], points(tp, te, pch = 16, col = "red"))
with(op1, plot(tp, te, type = "l"))
with(op1[op1$pval < 0.05,], points(tp, te, pch = 16, col = "red"))

}
\seealso{
\link{xmap}, \link{simplex}
}
