% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruic-package.R
\docType{package}
\name{ruic-package}
\alias{ruic}
\alias{ruic-package}
\title{ruic: Unified Information-theoretic Causality for R}
\description{
R implementation of unified information-theoretic causality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yutakaos/rUIC}
  \item Report bugs at \url{https://github.com/yutakaos/rUIC/issues}
}

}
\author{
\strong{Maintainer}: Yutaka Osada \email{ytosada@egmail.com} (\href{https://orcid.org/0000-0001-5967-194X}{ORCID})

}
\keyword{internal}
