% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruic-package.R
\docType{package}
\name{rUIC}
\alias{rUIC}
\title{Unified Information-theoretic Causality for R.}
\description{
The \pkg{rUIC} package is an implementation of UIC algorithms.
    UIC is based on the unified theory of convergent cross mapping (CCM) and
    transfer entropy (TE). The functions of this package have similar interface
    to \pkg{rEDM}.
}
\details{
This package has three basic functions (\code{xmap}, \code{simplex} and \code{uic})
    and two wrapper functions (\code{uic.optimal} and \code{uic.marginal}) to perform UIC.

\strong{Main Functions}: 
    \itemize{
        \item \code{\link{xmap}}    - cross mapping
        \item \code{\link{simplex}} - simplex projection
        \item \code{\link{uic}}     - unified information-theoretic causality (UIC)
        \item \code{\link{uic.optimal}}  - UIC using the optimal embedding dimension
        \item \code{\link{uic.marginal}} - UIC marginalizing embedding dimensions
    }
}
\author{
\strong{Maintainer}  : Yutaka Osada
    \strong{Contributors}: Masayuki Ushio
}
\keyword{package}
