% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nn.R
\name{set_nn}
\alias{set_nn}
\title{Set nn for xmap, simplex and uic functions in rUIC package}
\usage{
set_nn(nn, E)
}
\arguments{
\item{nn}{the number of nearest neighbors to use.
If \code{nn = "e+1"}, \code{nn} is set as \code{E} + 1.}

\item{E}{the embedding dimension to use for time-delay embedding.}
}
\value{
An integer vector.
}
\description{
\code{set_nn} returns an integer vector, which is needed for xmap, simplex and uic
functions in rUIC package.
}
\examples{
set_nn("e+1", 1:4)  # 2 3 4 5
set_nn(5, 1:4)      # 5 5 5 5
set_nn(3:6, 1:4)    # 3 4 5 6

}
