% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruic-package.R
\docType{package}
\name{rUIC-package}
\alias{rUIC}
\alias{rUIC-package}
\title{rUIC: Unified Information-theoretic Causality for R}
\description{
R implementation of unified information-theoretic causality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yutakaos/rUIC}
  \item Report bugs at \url{https://github.com/yutakaos/rUIC/issues}
}

}
\author{
\strong{Maintainer}: Yutaka Osada \email{ytosada@egmail.com} (\href{https://orcid.org/0000-0001-5967-194X}{ORCID})

Other contributors:
\itemize{
  \item Masayuki Ushio (\href{https://orcid.org/0000-0003-4831-7181}{ORCID}) [contributor]
}

}
