% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruic-package.R
\docType{package}
\name{rUIC}
\alias{rUIC}
\title{Unified Information-theoretic Causality for R.}
\description{
The \pkg{rUIC} package is a experimental implementation of UIC algorithms.
  UIC is based on the unified theory of convergent cross mapping (CCM) and
  transfer entropy (TE). The functions of this package have similar interface
  to \pkg{rEDM} because the UIC algorithm was developed based on that of CCM.
}
\details{
This package is divided into a set of main functions to perform UIC.
\strong{Main Functions}: 
  \itemize{
    \item \code{\link{xmap}        } - cross mapping
    \item \code{\link{simplex}     } - simplex projection
    \item \code{\link{uic}         } - unified infomation-theoretic causality
    \item \code{\link{marginal_uic}} - marginal unified infomation-theoretic causality
  }
}
\author{
\strong{Maintainer}  : Yutaka Osada
  \strong{Contributors}: Masayuki Ushio
}
\keyword{package}
