% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmap.R
\name{xmap}
\alias{xmap}
\title{Cross Mapping Prediction}
\usage{
xmap(
  block,
  lib = c(1, NROW(block)),
  pred = lib,
  group = NULL,
  lib_var = 1,
  tar_var = lib_var,
  cond_var = NULL,
  norm = 2,
  E = 1,
  tau = 1,
  tp = 0,
  nn = "e+1",
  exclusion_radius = NULL,
  epsilon = NULL,
  is_naive = FALSE,
  knn_method = c("KD", "BF")
)
}
\arguments{
\item{block}{a data.frame or matrix where each column is a time series.}

\item{lib}{a two-column matrix (or two-element vector) where each row specifies the
first and last indices of time series to use for attractor reconstruction.}

\item{pred}{a two-column matrix (or two-element vector) where each row specifies the
first and last indices of time series to use for model predictions.}

\item{group}{the name or column index of a group variable, which specifies different
groups of time series data.}

\item{lib_var}{the names or column indices of library variables.}

\item{tar_var}{the names or column indices of target variables.}

\item{cond_var}{the names or column indeices of condition variables, which are used as
library variables without time-delay embedding.}

\item{norm}{the power of Lp distance. Maximum distance is used if \code{norm} == Inf.}

\item{E}{the embedding dimension used for time-delay embedding. Must be an integer.}

\item{tau}{the time-lag used for time-delay embedding. Must be an integer.}

\item{tp}{the time index to predict. Must be an integer.}

\item{nn}{the number of nearest neighbors. Must be an integer or "e+1".
If \code{nn = "e+1"} or \code{nn = 0}, \code{nn} is set to \code{E} + 1.
If \code{nn < 0}, \code{nn} is set to the number of all data.}

\item{exclusion_radius}{the filtering to exclude nearest neighbors if their time index is within
exclusion radius.}

\item{epsilon}{the filtering to exclude nearest neighbors if their distance is farther
away than epsilon.}

\item{is_naive}{specifies whether naive estimator is used or not.}

\item{knn_method}{the method to search nearest neighbors.
The KD-tree ("KD") or brute-force ("BF") search can be used.}
}
\value{
A list or data.frame with model predictions, where each column is:
\tabular{ll}{
\code{data} \tab \code{:} data values used for model prediction \cr
\code{pred} \tab \code{:} predicted values \cr
\code{enn}  \tab \code{:} effective number of nearest neighbors \cr
\code{sqe}  \tab \code{:} squared errors \cr
}
}
\description{
\code{xmap} returns model predictions from given multiple time series using
cross mapping.
}
\examples{
# simulate logistic map
tl <- 400  # time length
x <- y <- rep(NA, tl)
x[1] <- 0.4
y[1] <- 0.2
for (t in 1:(tl - 1)) {  # causality : x -> y
    x[t+1] = x[t] * (3.8 - 3.8 * x[t] - 0.0 * y[t])
    y[t+1] = y[t] * (3.5 - 3.5 * y[t] - 0.1 * x[t])
}
block <- data.frame(t=1:tl, x=x, y=y)

# cross mapping
out0 <- xmap(block, lib_var="x", tar_var="y", E=2, tau=1, tp=-1)
out1 <- xmap(block, lib_var="y", tar_var="x", E=2, tau=1, tp=-1)
par(mfrow=c(2, 1))
with(out0, plot(data, pred))
with(out1, plot(data, pred))

}
